Extract duplicate Taylor series logic to shared utility
