+++ Math/Geometry/Trigonometry/Trig_Functions/cosine.py
@@ -2,19 +2,7 @@
 from typing import Union
 
 
-def factorial(n: int) -> int:
-    """Calculate factorial of n."""
-    if n < 0:
-        raise ValueError("Factorial not defined for negative numbers.")
-    if n > 1000:
-        raise ValueError(
-            "Factorial calculation limit exceeded (maximum allowed is 1000)."
-        )
-    result = 1
-    for i in range(1, n + 1):
-        result *= i
-    return result
-
+from Math.Geometry.Trigonometry.taylor_series import cosine_taylor
 
 def cosine(radians: Union[int, float]) -> float:
     """
@@ -26,17 +14,4 @@ def cosine(radians: Union[int, float]) -> float:
     Returns:
     float: The cosine of the angle.
     """
-    cos_value = 1
-    sign = 1
-    
-    # Calculate cosine using Taylor series: cos(x) = Σ((-1)ⁿ × x^(2n)) / (2n)!
-    for idx in range(2, 100, 2):
-        fact = factorial(idx)
-        
-        if sign % 2 == 0:
-            cos_value += radians**idx / fact
-        else:
-            cos_value -= radians**idx / fact
-        sign += 1
-    
-    return cos_value
+    return cosine_taylor(radians)
