+++ Math/Geometry/Trigonometry/Trig_Functions/sine.py
@@ -2,19 +2,7 @@
 from typing import Union
 
 
-def factorial(n: int) -> int:
-    """Calculate factorial of n."""
-    if n < 0:
-        raise ValueError("Factorial not defined for negative numbers.")
-    if n > 1000:
-        raise ValueError(
-            "Factorial calculation limit exceeded (maximum allowed is 1000)."
-        )
-    result = 1
-    for i in range(1, n + 1):
-        result *= i
-    return result
-
+from Math.Geometry.Trigonometry.taylor_series import sine_taylor
 
 def sine(radians: Union[int, float]) -> float:
     """
@@ -26,14 +14,4 @@ def sine(radians: Union[int, float]) -> float:
     Returns:
     float: The sine of the angle.
     """
-    sine_value = float(radians)
-    term = float(radians)
-    radians_sq = float(radians * radians)
-    
-    # Calculate sine using Taylor series iteratively:
-    # Next term = Previous term * (-x^2) / ((2n)(2n+1))
-    for idx in range(3, 100, 2):
-        term *= -radians_sq / ((idx - 1) * idx)
-        sine_value += term
-        
-    return sine_value
+    return sine_taylor(radians)
